# Eaas/openapi.yaml
openapi: 3.0.3
info:
  title: EaaS — Escrow as a Service API
  description: |
    API-first, Africa-native escrow platform. Enables secure transactions
    between buyers and sellers by holding funds in escrow until delivery is confirmed.
    
    ## Authentication
    - **Customer/Admin:** Bearer JWT token via `Authorization: Bearer <token>`
    - **Merchant API:** API Key via `X-API-Key: sk_live_xxxx` header
    - **Webhooks from PSP:** Verified via `X-Interswitch-Signature` HMAC header
    
    ## Idempotency
    All POST requests that create resources must include `X-Idempotency-Key` header.
    
    ## Rate Limiting
    - 100 requests/minute per API key
    - Rate limit headers returned: `X-RateLimit-Limit`, `X-RateLimit-Remaining`
    
  version: 1.0.0
  contact:
    name: EaaS Developer Support
    email: developers@eaas.africa
servers:
  - url: https://api.eaas.africa/api/v1
    description: Production
  - url: https://sandbox.eaas.africa/api/v1
    description: Sandbox
...
paths:
...
  /auth/register/customer:
    post:
      tags: [Auth]
      summary: Register a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/register/merchant:
    post:
      tags: [Auth]
      summary: Register a new merchant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MerchantRegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              example:
                status: "success"
                data:
                  accessToken: "eyJhbGci..."
                  refreshToken: "eyJhbGci..."
                  expiresIn: 3600
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ─── MERCHANTS ─────────────────────────────────────────────
  /merchants/kyc:
    post:
      tags: [Merchants]
      summary: Submit KYC business details
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KycSubmitRequest'
      responses:
        '200':
          description: KYC submitted successfully
  /merchants/kyc/documents:
    post:
      tags: [Merchants]
      summary: Upload KYC document (CAC, utility bill, government ID)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                documentType:
                  type: string
                  enum: [CAC_CERTIFICATE, UTILITY_BILL, GOVERNMENT_ID, BANK_STATEMENT]
      responses:
        '200':
          description: Document uploaded successfully
  /merchants/api-key/regenerate:
    post:
      tags: [Merchants]
      summary: Regenerate API key (invalidates old key immediately)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: New API key generated
          content:
            application/json:
              schema:
                type: string
  /merchants/webhook:
    put:
      tags: [Merchants]
      summary: Configure webhook URL
      security:
        - apiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                webhookUrl:
                  type: string
                  example: "https://mystore.com/webhooks/eaas"
      responses:
        '200':
          description: Webhook configured

  # ─── ESCROW ────────────────────────────────────────────────
...
  /users/me/data-export:
    get:
      tags: [User Data Rights]
      summary: Export my data
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User data export
          content:
            application/json:
              schema:
                type: string
  /users/me:
    delete:
      tags: [User Data Rights]
      summary: Delete my account (anonymize)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account scheduled for deletion
  /users/me/marketing-consent:
    put:
      tags: [User Data Rights]
      summary: Update marketing consent
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: consent
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: Marketing consent updated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          example:
            status: "error"
            message: "Validation failed"
            errors:
              - field: "amount"
                message: "Amount must be greater than zero"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          example:
            status: "error"
            message: "Unauthorized"
...
    Conflict:
      description: State conflict
      content:
        application/json:
          example:
            status: "error"
            message: "Cannot transition from CONFIRMED to DISPUTED"
  schemas:
    ApiResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
        message:
          type: string
        data:
          type: object
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    CustomerRegisterRequest:
      type: object
      required: [fullName, email, phone, password, termsAccepted, dataProcessingConsent]
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 8
        termsAccepted:
          type: boolean
        dataProcessingConsent:
          type: boolean
        marketingConsent:
          type: boolean
    MerchantRegisterRequest:
      type: object
      required: [fullName, email, phone, password, businessName, businessRegNumber, bankAccountNumber, bankCode, bankName, bvn, termsAccepted, dataProcessingConsent]
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 8
        businessName:
          type: string
        businessType:
          type: string
        businessRegNumber:
          type: string
        bankAccountNumber:
          type: string
        bankCode:
          type: string
        bankName:
          type: string
        bvn:
          type: string
          pattern: "^\\d{11}$"
        termsAccepted:
          type: boolean
        dataProcessingConsent:
          type: boolean
        marketingConsent:
          type: boolean
    KycSubmitRequest:
      type: object
      required: [businessName, bankAccountNumber, bankCode, bankName, bvn]
      properties:
        businessName:
          type: string
        businessType:
          type: string
        businessRegNumber:
          type: string
        bankAccountNumber:
          type: string
          pattern: "^\\d{10}$"
        bankCode:
          type: string
        bankName:
          type: string
        bvn:
          type: string
          pattern: "^\\d{11}$"
...